Extract type expressions with all tests remaining working
